Feat/lang dropdown sreya (#24)
* Made a functioning dropdown

Created the functional aspects of the dropdown: when the v button is clicked, a table of options is created; the clicking of different languages is saved into a state variable; and the component that shows the selected language gets rerendered based on the state variable. Also, I saved the state variables into a context that hopefully allows other components to be able to reference the language selection : )

* Moved the LanguageContext to a higher page

In the previous commit, I created a context to store the "language" state variable. In this commit, I moved the context to the index page, so all the child components of the index will get access to the context: the currently selected language and a function to change the language (I put the function for the dropdown component to access TT)

* Created hover function and close when clicking out of div function

* Added some animation and styles

Made the dropdown rollout animation!

* Added icons and dividers to dropdown

* Added more styling details!

* Finalized dropdown!! (added original language names and fixed context values)

Made the language names to be written in the corresponding language. Made the returned context values of each option to match the names of json's (en, zh, ko, vi, hmn, and es) :)

* Added i18 features in dropdown

* Added i18n compatibility + Condensed dropdown css & code

I deleted repetitive code in dropDown.jsx using object mapping & created a i18n.config. I also left an example, compatible to my additions in dropDown.jsx, in index.jsx on how I am imagining i18n would work

* Replaced imgs with React Icons + Cleaned up dropDown css

* language change simplified

* Added mobile structure to dropdown

* Added header collapse/expand functionality

* Mobile dropdown toggle fixed

* Reorganized dropDown css file

* mobile dropdown simplified

---------

Co-authored-by: JayJ104 <123anjalijain@gmail.com>