Feat/header-nandhana (#16)
* Made a functioning dropdown

Created the functional aspects of the dropdown: when the v button is clicked, a table of options is created; the clicking of different languages is saved into a state variable; and the component that shows the selected language gets rerendered based on the state variable. Also, I saved the state variables into a context that hopefully allows other components to be able to reference the language selection : )

* Moved the LanguageContext to a higher page

In the previous commit, I created a context to store the "language" state variable. In this commit, I moved the context to the index page, so all the child components of the index will get access to the context: the currently selected language and a function to change the language (I put the function for the dropdown component to access TT)

* Created hover function and close when clicking out of div function

* added basic structure of page links

* added underline on hover. attempting bold on current page

* fixed fonts

* Added some animation and styles

Made the dropdown rollout animation!

* added image

* attempting useState for highlight on active

* Added icons and dividers to dropdown

* Added more styling details!

* Finalized dropdown!! (added original language names and fixed context values)

Made the language names to be written in the corresponding language. Made the returned context values of each option to match the names of json's (en, zh, ko, vi, hmn, and es) :)

* minor spacing fixes + added border on active page

* added header folder for image

* fixes- highlight, underline issue,

* working on internationalization

* trying to add internationalization

* Revert "Merge branch 'feat/header-nandhana' of https://github.com/include-davis/paul-homm into feat/header-nandhana"

This reverts commit 463725af7fe819680d2ea8b09ff6fe65e90765ff, reversing
changes made to 3c90ac6c758ee2ab7cf5363b54b2f781601c3a55.

* Added i18 features in dropdown

* starting to implement mobile header

* working on mobile design

* added minor comment for clarity

* Added i18n compatibility + Condensed dropdown css & code

I deleted repetitive code in dropDown.jsx using object mapping & created a i18n.config. I also left an example, compatible to my additions in dropDown.jsx, in index.jsx on how I am imagining i18n would work

* Replaced imgs with React Icons + Cleaned up dropDown css

* header link onclick fix

* language change simplified

* Update en.json

* language dropdown positioning

* Update footer.module.scss

* implemented tablet mixins

* implemented mobile mixins

* Added mobile structure to dropdown

* Added header collapse/expand functionality

* Mobile dropdown toggle fixed

* Reorganized dropDown css file

* increased z index of hamburger menu dropdown

* implemented fixes to link highliting

Corrects the errors where the links wouldn't span the entire menu. Only thing left is to fix the background on active pages to achieve the right dimensions & possibly add in a line transformation on hover.

* fixed mobile dropdown (z index, highlights)

* mobile dropdown simplified

* language dropdown position fixed

---------

Co-authored-by: tiramisuuuuuuu <ssmathew@ucdavis.edu>
Co-authored-by: Jay <116294517+JayJ104@users.noreply.github.com>
Co-authored-by: JayJ104 <123anjalijain@gmail.com>
Co-authored-by: Samantha Iraheta <sairaheta1@gmail.com>